+++ sdk/api/v1alpha1/types.go
@@ -10,11 +10,21 @@ const (
 
 	// ProviderAzure represents the new Azure provider which is implemented with Cluster API.
 	ProviderAzure Provider = "azure"
+
+	// ProviderVsphere represents the new Vsphere provider which is implemented with Cluster API.
+	ProviderVsphere Provider = "vsphere"
+
+	// ProviderCloudDirector represents the new VCD provider which is implemented with Cluster API.
+	ProviderCloudDirector Provider = "cloud-director"
 )
 
 var SupportedProviders = []Provider{
 	ProviderAws,
 	ProviderAzure,
+	ProviderVsphere,
+
+	// Not yet supported:
+	// ProviderCloudDirector,
 }
 
 type Provider string