+++ sdk/CHANGELOG.md
@@ -5,9 +5,25 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-
-
 ## [Unreleased]
+
+## [0.8.0] - 2024-10-15
+
+### Added
+
+- Add Cloud Director provider to the list of supported providers.
+
+### Changed
+
+- Made the `getProviderDirectory` function more dynamic and less hardcoded for most cases.
+
+## [0.7.0] - 2024-10-15
+
+### Added
+
+- Add VSphere provider to the list of supported providers.
+
+## [0.6.0] - 2024-07-26
 
 ### Changed